Milestone 2!

See merge request kfg1009/bullet-zone-coruscant-24-25!20

+++ server/src/main/java/edu/unh/cs/cs619/bulletzone/datalayer/account/BankAccount.java
@@ -1,14 +1,13 @@
 package edu.unh.cs.cs619.bulletzone.datalayer.account;
 
 import edu.unh.cs.cs619.bulletzone.datalayer.permission.OwnableEntity;
-import edu.unh.cs.cs619.bulletzone.datalayer.core.EntityRecord;
 
 public class BankAccount extends OwnableEntity {
     protected double balance;
 
     public double getBalance() { return balance; }
 
-    public BankAccount(int userId) {
+    public BankAccount(long userId) {
         super(new BankAccountRecord());
         this.balance = 1000.0; // Initial balance of 1000 credits
     }
